removed all repositories from previous tasks
