+++ src/main/java/co/inventrosoft/hibernate_task/model/Tournament.java
@@ -12,6 +12,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
+
 @Entity
 @Data
 @AllArgsConstructor
