+++ src/main/java/co/inventrosoft/hibernate_task/repository/TournamentRepository.java
@@ -1,6 +1,12 @@
 package co.inventrosoft.hibernate_task.repository;
 
 
-public interface TournamentRepository {
-    int create();
+import co.inventrosoft.hibernate_task.model.Tournament;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface TournamentRepository extends JpaRepository<Tournament, Integer> {
+
 }
