+++ src/main/java/co/inventrosoft/hibernate_task/service/MatchService.java
@@ -3,7 +3,7 @@ package co.inventrosoft.hibernate_task.service;
 import co.inventrosoft.hibernate_task.console.MatchResult;
 import co.inventrosoft.hibernate_task.model.Match;
 import co.inventrosoft.hibernate_task.model.Team;
-import co.inventrosoft.hibernate_task.repository.hibernate.MatchRepository;
+import co.inventrosoft.hibernate_task.repository.MatchRepository;
 import org.springframework.stereotype.Service;
 
 import java.io.FileNotFoundException;
