+++ src/main/java/co/inventrosoft/hibernate_task/service/TeamService.java
@@ -2,7 +2,7 @@ package co.inventrosoft.hibernate_task.service;
 
 import co.inventrosoft.hibernate_task.console.MatchResult;
 import co.inventrosoft.hibernate_task.model.Team;
-import co.inventrosoft.hibernate_task.repository.hibernate.TeamRepository;
+import co.inventrosoft.hibernate_task.repository.TeamRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
